refactor(tests): moved to separate files & added conftest.py for fixtures
